+++ xs/include/xs/mir.h
@@ -89,6 +89,7 @@ typedef enum
   XS_MIR_INSTRUCTION_LE_I32,
   XS_MIR_INSTRUCTION_GT_I32,
   XS_MIR_INSTRUCTION_GE_I32,
+  XS_MIR_INSTRUCTION_NOT_BOOL,
   XS_MIR_INSTRUCTION_LOAD,
   XS_MIR_INSTRUCTION_STORE,
 } XsMirInstructionKind;
@@ -209,6 +210,7 @@ XsMirStatus xs_mir_block_gt_i32(XsMirBlock *block, XsMirValueId left, XsMirValue
                                 XsMirError *error);
 XsMirStatus xs_mir_block_ge_i32(XsMirBlock *block, XsMirValueId left, XsMirValueId right, XsMirValueId *result,
                                 XsMirError *error);
+XsMirStatus xs_mir_block_not_bool(XsMirBlock *block, XsMirValueId operand, XsMirValueId *result, XsMirError *error);
 XsMirStatus xs_mir_block_add_load(XsMirBlock *block, const XsMirPlace *place, XsMirType result_type,
                                   XsMirValueId *result, XsMirError *error);
 XsMirStatus xs_mir_block_add_store(XsMirBlock *block, const XsMirPlace *place, XsMirValueId value, XsMirError *error);