+++ xs/sources/mir/xlil_lowering.c
@@ -73,6 +73,13 @@ static XsMirStatus lower_instruction(const XsMirInstruction *instruction, XsLilB
         xs_lil_block_add_const_bool(block, instruction->immediate_i64 != 0, &values[instruction->result], &lil_error);
     return map_lil_status(status, &lil_error, error);
   }
+  case XS_MIR_INSTRUCTION_NOT_BOOL:
+  {
+    XsLilError lil_error = {0};
+    XsLilStatus status =
+        xs_lil_block_not_bool(block, values[instruction->operand_left], &values[instruction->result], &lil_error);
+    return map_lil_status(status, &lil_error, error);
+  }
   case XS_MIR_INSTRUCTION_ADD_I64:
   {
     XsLilError lil_error = {0};